Move error types of storage to `storage::error`
